fix: Resolve backend startup errors and Swagger API visibility
- Switch Redisson dependency to bypass Spring Boot 4.1.0 compatibility issue
- Fix RedisConfig syntax errors
- Resolve Flyway version conflict (rename V2 to V8)
- Add V9 migration to align shift_swap_request schema with JPA entity
- Update V7 migration with IF NOT EXISTS to prevent index creation failure
- Fix JPQL query in ShiftAssignmentRepository (remove non-existent 'status' field)
- Update PayrollAggregation DTOs to match JPQL SUM return type (BigDecimal)
- Update DashboardService to correctly cast BigDecimal to double
- Change spring.jpa.hibernate.ddl-auto to 'update' to auto-create missing tables (like user_device_tokens)- Add Swagger tags to AttendanceController

+++ shiftsync-backend/src/main/java/com/shiftsync/payroll/dto/PayrollChartAggregation.java
@@ -3,5 +3,5 @@ package com.shiftsync.payroll.dto;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record PayrollChartAggregation(LocalDate startDate, LocalDate endDate, Double totalHours, BigDecimal totalAmount) {
+public record PayrollChartAggregation(LocalDate startDate, LocalDate endDate, BigDecimal totalHours, BigDecimal totalAmount) {
 }
